* More tidying up in the demuxer library.
git-svn-id: https://svn.xiph.org/trunk/oggdsf@7325 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ src/lib/core/directshow/dsfOggDemux/StreamHeaders.cpp
@@ -53,7 +53,7 @@ StampedOggPacket* StreamHeaders::getPacket(unsigned long inPacketNo) {
 }
 
 unsigned long StreamHeaders::numPackets() {
-	return mPacketList.size();
+	return (unsigned long)mPacketList.size();
 }
 
 
@@ -62,7 +62,7 @@ bool StreamHeaders::addPacket(StampedOggPacket* inPacket) {
 		mPacketList.push_back(inPacket);
 
 		StampedOggPacket* loc = NULL;
-		for (int i = 0; i < mPacketList.size(); i++) {
+		for (size_t i = 0; i < mPacketList.size(); i++) {
 			loc = mPacketList[i];
 		}
